requirements.txt
-i https://pypi.org/simple
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
django-filter==2.1.0
django-rest-swagger==2.2.0
django==2.2.1
djangorestframework==3.9.3
idna==2.8
itypes==1.1.0
jinja2==2.10.1
markupsafe==1.1.1
openapi-codec==1.3.2
pytz==2019.1
requests==2.21.0
simplejson==3.16.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.3
